% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BINSEG.R
\name{BINSEG}
\alias{BINSEG}
\title{Performs Binary Segmentation on a data matrix containing a log-ratio
numeric vector.}
\usage{
BINSEG(data.dt, num_segments)
}
\arguments{
\item{data.dt}{A numeric data data containing a log-ratio numeric Vector}

\item{num_segments}{The number of segments}
}
\value{
Returns the number of segments in the model. This function is not
currently working
}
\description{
This function computes binary segmentation on a data table with a logration
column.
}
\details{
THIS FUNCTION IS NOT CURRENTLY WORKING!
}
\examples{
data(neuroblastoma, package="neuroblastoma")
library(data.table)
proifle_id <- "4"
chromosome_num <- "2"
nb.dt <- data.table(neuroblastoma$profiles)
data.dt <- nb.dt[profile.id==proifle_id & chromosome==chromosome_num]
num_segs <- 20
my.binseg.result <- BINSEG(data.dt, num_segs)

}
