% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeckOfCardS4.R
\docType{class}
\name{BootOfDecks4}
\alias{BootOfDecks4}
\title{Boot of Decks Class}
\description{
A class representing a collection (boot) of multiple decks of cards.
}
\section{Slots}{

\describe{
\item{\code{decks}}{A SetOfCards object.}
}}

