% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_dice.R
\name{coin_flip}
\alias{coin_flip}
\title{Coin Flip}
\usage{
coin_flip(flips = 1, coins = 1)
}
\arguments{
\item{flips}{The number of times to flip the coin.}

\item{coins}{The number of coins to flip.}
}
\value{
A vector of integers representing the results of the coin flips.
}
\description{
This function simulates flipping a coin.
}
\examples{

coin_flip(1, 1)
coin_flip(10, 1)
coin_flip(1, 10)

}
