% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeckOfCardS4.R
\name{createBootDeck}
\alias{createBootDeck}
\title{Function to create a Boot of S4 DeckOfCards4.}
\usage{
createBootDeck(num_decks = 2, use_jokers = FALSE)
}
\arguments{
\item{use_jokers}{A logical value indicating whether to include jokers in the deck.}
}
\value{
A data frame representing a standard deck of playing cards.
}
\description{
Function to create a Boot of S4 DeckOfCards4.
}
\examples{
createBootDeck()
}
