% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeckOfCardS4.R
\name{createDeck}
\alias{createDeck}
\title{Function to create a standard deck dataframe.}
\usage{
createDeck(use_jokers = FALSE)
}
\arguments{
\item{use_jokers}{A logical value indicating whether to include jokers in the deck.}
}
\value{
A data frame representing a standard deck of playing cards.
}
\description{
Function to create a standard deck dataframe.
}
\examples{
createDeck()
}
