% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeckOfCards.R
\name{DeckOfCards}
\alias{DeckOfCards}
\title{DeckOfCards}
\description{
This R6 class simulates a deck of cards that allows drawing and reshuffling.
}
\details{
Deck of Cards

This R6 class simulates a deck of cards that allows drawing and reshuffling.
}
\examples{
# Create a new deck
my_deck <- DeckOfCards$new()
# Shuffle the deck
my_deck$reshuffle()
# Draw cards
drawn_cards <- my_deck$draw(5)
# Print drawn cards
print(my_deck$drawn)
# Draw more cards
drawn_cards <- my_deck$draw(5)
# Print remaining cards
print(my_deck$deck)
# Reshuffle the deck
my_deck$reshuffle()


}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{deck}}{A data frame representing the deck of cards.}

\item{\code{drawn}}{A data frame representing the drawn cards.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DeckOfCards-new}{\code{DeckOfCards$new()}}
\item \href{#method-DeckOfCards-draw}{\code{DeckOfCards$draw()}}
\item \href{#method-DeckOfCards-reshuffle}{\code{DeckOfCards$reshuffle()}}
\item \href{#method-DeckOfCards-clone}{\code{DeckOfCards$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeckOfCards-new"></a>}}
\if{latex}{\out{\hypertarget{method-DeckOfCards-new}{}}}
\subsection{Method \code{new()}}{
Initialize the deck of cards.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeckOfCards$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a data frame representing the deck of cards.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeckOfCards-draw"></a>}}
\if{latex}{\out{\hypertarget{method-DeckOfCards-draw}{}}}
\subsection{Method \code{draw()}}{
Draw a specified number of cards from the deck.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeckOfCards$draw(draws = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{draws}}{The number of cards to draw from the deck.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame representing the drawn cards.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeckOfCards-reshuffle"></a>}}
\if{latex}{\out{\hypertarget{method-DeckOfCards-reshuffle}{}}}
\subsection{Method \code{reshuffle()}}{
Reshuffle the deck of cards.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeckOfCards$reshuffle()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a data frame representing the reshuffled deck of cards.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeckOfCards-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DeckOfCards-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeckOfCards$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
