% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeckOfCardS4.R
\docType{class}
\name{DeckOfCards4}
\alias{DeckOfCards4}
\title{Deck of Cards Class}
\description{
A class inheriting from SetOfCards, representing a standard deck of playing cards.
}
\section{Slots}{

\describe{
\item{\code{setofcards}}{A data frame containing a deck of cards and their suits.}
}}

