% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeckOfCardS4.R
\docType{class}
\name{DiscardPile4}
\alias{DiscardPile4}
\title{Discard S4 Pile Class}
\description{
A class representing the pile of cards that have been discarded.
}
\section{Slots}{

\describe{
\item{\code{discard}}{A SetOfCards object representing the discarded cards.}
}}

