% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_percent_miss.R,
%   R/plot.features_percent_miss.R
\name{features_percent_miss}
\alias{features_percent_miss}
\alias{plot.features_percent_miss}
\title{Features Percent Missing}
\usage{
features_percent_miss(data, percent_miss = 0, na_list = NULL)

\method{plot}{features_percent_miss}(x, add.label = TRUE)
}
\arguments{
\item{data}{a data frame or tibble or something that could become one}

\item{percent_miss}{numeric between 0 an 1 will filter out any columns that are missing more than this value}

\item{na_list}{list of elements to allow SumNa to think of as NA}

\item{x}{An object of class 'features_percent_miss' returned by the
\code{\link{features_percent_miss}} function.}

\item{add.label}{logical add percent labels to graph, defaults to TRUE}
}
\value{
a tibble (data frame) with columns: feature, SumNa (sum of NA entries), SumComp, PctNa, PctComp

A ggplot object representing the bar plot.
}
\description{
Count the number of missing elements of each column in a data frame.
Allows for threshold setting to easily find columns with higher percent missing data.
Allows users to supply list of things that might also be be considered missing.

This function creates a bar plot showing the percentage of complete values
for each feature in a data frame or tibble.
}
\examples{
data(input_parameters_raw)

features_percent_miss(input_parameters_raw,  na_list = c(""," ","NA"))

data(input_parameters_raw)

input_parameters_raw |>
  features_percent_miss(na_list = c(""," ","NA")) |>
  plot()

}
