\name{dynprog_interface}
\docType{package}
\title{
  Dynamic Progamming Algorithm
}
\alias{
  dynprog_interface
}
\description{
  TODO
}
\usage{
dynprog_interface(data_mat, clusters)
}
\arguments{
\item{data_mat}{A numeric data matrix (nxp)}

\item{clusters}{The number of clusters}
}
\value{
TODO 
}
\examples{
  TODO
}