% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_at_random.R
\name{missing_at_random}
\alias{missing_at_random}
\title{Missing data at random}
\usage{
missing_at_random(data, prob = runif(1, 1/5, 1/3))
}
\arguments{
\item{data}{A dataset}

\item{prob}{The probability of a missing value being assigned to a cell}
}
\value{
A dataset with missing values randomly assigned with the given probability
}
\description{
This function will randomly assign missing values to a dataset
}
\examples{
mtcars_miss_at_random <- missing_at_random(mtcars)

}
