% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missmap.R
\name{missmap}
\alias{missmap}
\title{Missingness Map}
\usage{
missmap(df, na_list = NULL, row_order = FALSE)
}
\arguments{
\item{df}{A dataframe}

\item{na_list}{list of elements to allow SumNa to think of as NA}

\item{row_order}{logical value to determine if the rows should be ordered by the number of missing values}
}
\value{
A ggplot object
}
\description{
This function creates a missingness map for a dataframe. The map shows the missingness of each cell in the dataframe.
}
\examples{
data <- data.frame(
col1 = c(1, NA, 3, "NA"),
col2 = c("A", "B", "C", "D")
)
missmap(data)

mtcars |>
  missing_at_random() |>
  missmap()


}
