% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiplication.R
\name{multiplication}
\alias{multiplication}
\title{Multiply two numbers}
\usage{
multiplication(x, y)
}
\arguments{
\item{x}{First number}

\item{y}{Second number}
}
\value{
Product of x and y
}
\description{
Multiply two numbers
}
\examples{
multiplication(3, 5)

}
