% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeckOfCardS4.R
\docType{class}
\name{PlayerHand4}
\alias{PlayerHand4}
\title{Player Hand Class}
\description{
A class representing a player's hand of cards.
}
\section{Slots}{

\describe{
\item{\code{name}}{A character string representing the player's name.}

\item{\code{hand}}{A SetOfCards object representing the player's hand.}
}}

