\name{rpackage2-package}
\alias{rpackage2-package}
\alias{rpackage2}
\docType{package}
\title{
  Clustering Algorithms (Kmeans and Gaussian Mixture Model)
}
\description{
  This package is a project for my Unsupervised Learning course.This package 
  contains segmentations alogrithms for Binary Segmentation and Dynamic 
  Programming.
  Dynamic Programming is coded using c++ and rcpp and Binary Segmentation 
  is coded using R and is not working. 
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Jacob Kaufman

Maintainer: Jacob Kaufman <jmk478@nau.edu>
}
\keyword{ segmentation }