% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeckOfCardS4.R
\docType{class}
\name{SetOfCards}
\alias{SetOfCards}
\title{Set of Cards Class}
\description{
A class representing a set of playing cards.
}
\section{Slots}{

\describe{
\item{\code{setofcards}}{A data frame containing a data frame of a set of cards and their suits.}
}}

