% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeckOfCardS4.R
\name{shuffle}
\alias{shuffle}
\title{Shuffle a Set of Cards}
\usage{
shuffle(object)
}
\arguments{
\item{object}{A SetOfCards object.}
}
\value{
The shuffled SetOfCards object.
}
\description{
Randomly permutes the cards in the set.
}
\examples{
my_deck <- new("DeckOfCards4", setofcards = createDeck())
my_deck <- shuffle(my_deck)
}
