% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeckOfCardS4.R
\name{solitaire}
\alias{solitaire}
\title{Sort a Set of Cards in Solitaire Order}
\usage{
solitaire(object)
}
\arguments{
\item{object}{A SetOfCards object.}
}
\value{
The SetOfCards object sorted in solitaire order.
}
\description{
Sorts the cards in the set by suit and rank in a specific order.
}
\examples{
my_deck <- new("DeckOfCards4", setofcards = createDeck())
my_deck_sorted <- solitaire(my_deck)
my_deck_sorted
}
