% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumNa.R
\name{SumNa}
\alias{SumNa}
\title{Count the number of missing elements of a vector}
\usage{
SumNa(col, na_list = NULL)
}
\arguments{
\item{col}{column or vector of data}

\item{na_list}{list of elements to allow SumNa to think of as NA}
}
\value{
numeric value representing number of missing values
}
\description{
Count the number of missing elements of a vector,
allows user to define what should count as \code{NA} in R
The function is vectorized.
}
\examples{
data <- data.frame(
 col1 = c(1, NA, 3, "NA"),
 col2 = c("A", "B", "C", "D")
 )
SumNa(data$col1)
SumNa(data$col1, na_list = c("NA"))

}
