% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Addition.R
\name{addition}
\alias{addition}
\title{Add two numbers}
\usage{
addition(x, y)
}
\arguments{
\item{x}{First number}

\item{y}{Second number}
}
\value{
Sum of x and y
}
\description{
Add two numbers
}
\examples{
addition(3, 5)

}
