% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeckOfCards.R
\name{BootOfDecks}
\alias{BootOfDecks}
\title{Boot of Decks}
\description{
Boot of Decks

Boot of Decks
}
\details{
This R6 class simulates a boot of decks that allows drawing and reshuffling.
}
\examples{
# Create a boot of decks
my_boot <- BootOfDecks$new(number_of_decks = 2)
# Draw cards from the boot
drawn_cards <- my_boot$draw(5)
# Print drawn cards
print(drawn_cards)
# Draw more cards
drawn_cards <- my_boot$draw(5)
# Print remaining cards
print(my_boot$drawn)
# Reshuffle the boot
my_boot$reshuffle()


}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{decks}}{A list of DeckOfCards objects.}

\item{\code{drawn}}{A data frame representing the drawn cards.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BootOfDecks-new}{\code{BootOfDecks$new()}}
\item \href{#method-BootOfDecks-reshuffle}{\code{BootOfDecks$reshuffle()}}
\item \href{#method-BootOfDecks-draw}{\code{BootOfDecks$draw()}}
\item \href{#method-BootOfDecks-clone}{\code{BootOfDecks$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BootOfDecks-new"></a>}}
\if{latex}{\out{\hypertarget{method-BootOfDecks-new}{}}}
\subsection{Method \code{new()}}{
Initialize the boot of decks.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BootOfDecks$new(number_of_decks)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{number_of_decks}}{The number of decks in the boot.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of DeckOfCards objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BootOfDecks-reshuffle"></a>}}
\if{latex}{\out{\hypertarget{method-BootOfDecks-reshuffle}{}}}
\subsection{Method \code{reshuffle()}}{
Reshuffle the boot of decks.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BootOfDecks$reshuffle()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data frame representing the reshuffled boot of decks.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BootOfDecks-draw"></a>}}
\if{latex}{\out{\hypertarget{method-BootOfDecks-draw}{}}}
\subsection{Method \code{draw()}}{
Draw a specified number of cards from the boot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BootOfDecks$draw(draws = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{draws}}{The number of cards to draw from the boot.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame representing the drawn cards.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BootOfDecks-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BootOfDecks-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BootOfDecks$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
