% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeckOfCards.R
\name{DeckOfCards}
\alias{DeckOfCards}
\title{DeckOfCards}
\description{
This R6 class simulates a deck of cards that allows drawing and reshuffling.
}
\details{
Deck of Cards

This R6 class simulates a deck of cards that allows drawing and reshuffling.
}
\examples{
# Create a new deck
my_deck <- DeckOfCards$new()
# Shuffle the deck
my_deck$reshuffle()
# Draw cards
drawn_cards1 <- my_deck$draw(5)
# Print drawn cards
print(my_deck$drawn)
# should be the same
drawn_cards1 == my_deck$drawn
# Draw more cards
drawn_cards2 <- my_deck$draw(3)
# should be the same
rbind(drawn_cards1, drawn_cards2) == my_deck$drawn
# Print drawn cards
print(my_deck$drawn)
# Check number of cards
nrow(rbind(drawn_cards1, drawn_cards2)) + nrow(my_deck$deck) == 52
# Reshuffle the deck
my_deck$reshuffle()
nrow(my_deck$deck) == 52
nrow(my_deck$drawn) == 0
# draw 40 cards
drawn_cards3 <- my_deck$draw(40)
# sort the remaining cards
my_deck$solitaire("deck") |> print()


}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{deck}}{A data frame representing the deck of cards.}

\item{\code{drawn}}{A data frame representing the drawn cards.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DeckOfCards-new}{\code{DeckOfCards$new()}}
\item \href{#method-DeckOfCards-draw}{\code{DeckOfCards$draw()}}
\item \href{#method-DeckOfCards-reshuffle}{\code{DeckOfCards$reshuffle()}}
\item \href{#method-DeckOfCards-solitaire}{\code{DeckOfCards$solitaire()}}
\item \href{#method-DeckOfCards-clone}{\code{DeckOfCards$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeckOfCards-new"></a>}}
\if{latex}{\out{\hypertarget{method-DeckOfCards-new}{}}}
\subsection{Method \code{new()}}{
Initialize the deck of cards.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeckOfCards$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a data frame representing the deck of cards.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeckOfCards-draw"></a>}}
\if{latex}{\out{\hypertarget{method-DeckOfCards-draw}{}}}
\subsection{Method \code{draw()}}{
Draw a specified number of cards from the deck.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeckOfCards$draw(draws = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{draws}}{The number of cards to draw from the deck.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame representing the drawn cards.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeckOfCards-reshuffle"></a>}}
\if{latex}{\out{\hypertarget{method-DeckOfCards-reshuffle}{}}}
\subsection{Method \code{reshuffle()}}{
Reshuffle the deck of cards.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeckOfCards$reshuffle()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a data frame representing the reshuffled deck of cards.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeckOfCards-solitaire"></a>}}
\if{latex}{\out{\hypertarget{method-DeckOfCards-solitaire}{}}}
\subsection{Method \code{solitaire()}}{
sort the deck of cards.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeckOfCards$solitaire(hand = "deck")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hand}}{either the \code{deck} or the \code{drawn} hand.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a data frame representing the sorted deck of cards.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeckOfCards-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DeckOfCards-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeckOfCards$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
