\name{rpackage2-package}
\alias{rpackage2-package}
\alias{rpackage2}
\docType{package}
\title{
  Clustering Algorithms (Kmeans and Gaussian Mixture Model)
}
\description{
  This package is a project for my Unsupervised Learning course.This package 
  contains clustering alogrithms for Kmeans and Gaussian Mixture Models (GMM). 
  Kmeans is coded using c++ and rcpp and GMM is coded using R. 
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Jacob Kaufman

Maintainer: Jacob Kaufman <jmk478@nau.edu>
}
\keyword{ clustering }